# openapi.yaml
openapi: 3.1.3
info:
  title: Ludus Extension API
  version: "1.0"
  description: API for managing scenarios, data, and topology in Ludus CTFd extension.
servers:
  - url: http://127.0.0.1:5000
components:
  securitySchemes:
    ApiKeyAuth: 
      type: apiKey 
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
      required:
        - error
security:
  - ApiKeyAuth: []
paths:
...
  /ctfd/scenario:
    get:
      summary: Get scenario(s)
      description: Get a scenario by ID (as base64) or list all scenarios.
      tags:
        - CTFd Scenario
      parameters:
        - in: query
          name: scenarioId
          schema:
            type: string
          required: false
          description: Scenario ID
          example: "GHI123"
      responses:
        '200':
          description: Scenario(s) found
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      scenarioId:
                        type: string
                        pattern: "^[a-zA-Z0-9]{6}$"
                      scenarioName:
                        type: string
                      scenarioFile:
                        type: string
                        description: base64-encoded file
                      createdAt:
                        type: string
                        description: File creation timestamp
                  - type: array
                    items:
                      type: object
                      properties:
                        scenarioId:
                          type: string
                          pattern: "^[a-zA-Z0-9]{6}$"
                        scenarioName:
                          type: string
                        createdAt:
                          type: string
                          description: File creation timestamp
        '400':
          description: Bad Request
        '404':
          description: Not Found

    put:
      summary: Create or update a scenario
      description: Upload a .zip file to create or update a scenario.
      tags:
        - CTFd Scenario
      parameters:
        - in: query
          name: scenarioId
          schema:
            type: string
            pattern: "^[a-zA-Z0-9]{6}$"
          required: false
          description: Scenario ID (if updating)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Uploaded successfully
        '400':
          description: Bad Request
        '404':
          description: Not Found

    delete:
      summary: Delete a scenario
      tags:
        - CTFd Scenario
      parameters:
        - in: query
          name: scenarioId
          schema:
            type: string
            pattern: "^[a-zA-Z0-9]{6}$"
          required: true
          description: Scenario ID
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Bad Request
        '404':
          description: Not Found

  # Topology
  /topology:
    get:
      summary: Get topology
      tags:
        - Topology
      parameters:
        - in: query
          name: topologyId
          schema:
            type: string
          required: false
          description: Topology ID
          example: "XYZ789"
      responses:
        '200':
          description: Topology found
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      topologyId:
                        type: string
                        pattern: "^[a-zA-Z0-9]{6}$"
                      topologyName:
                        type: string
                      topologyFile:
                        type: string
                        description: base64-encoded file
                      createdAt:
                        type: string
                        description: File creation timestamp
                  - type: array
                    items:
                      type: object
                      properties:
                        topologyId:
                          type: string
                          pattern: "^[a-zA-Z0-9]{6}$"
                        topologyName:
                          type: string
                        createdAt:
                          type: string
                          description: File creation timestamp
        '400':
          description: Bad Request
        '404':
          description: Not Found

    put:
      summary: Create or update topology
      tags:
        - Topology
      parameters:
        - in: query
          name: topologyId
          schema:
            type: string
          required: false
          description: Topology ID (if updating)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Uploaded successfully
        '400':
          description: Bad Request
        '404':
          description: Not Found

    delete:
      summary: Delete topology
      tags:
        - Topology
      parameters:
        - in: query
          name: topologyId
          schema:
            type: string
          required: true
          description: Topology ID
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Bad Request
        '404':
          description: Not Found
  
  # Pool
  /pool:
    post:
      summary: Create a new pool
      description: Create a new pool with topology, users, and optional teams
      tags:
        - Pool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: ["SHARED", "INDIVIDUAL"]
                  example: "SHARED"
                topologyId:
                  type: string
                  example: "H4tCgb"
                usersAndTeams:
                  type: array
                  items:
                    type: object
                    properties:
                      user:
                        type: string
                        example: "Alice Dan Mayers"
                      team:
                        type: string
                        example: "Red Team"
                      mainUserId:
                        type: string
                        example: "instructor1"
                        description: "Required for SHARED pools only"
                    required:
                      - user
                    additionalProperties: false
                note:
                  type: string
                  example: "Training session"
              required:
                - type
                - topologyId
                - usersAndTeams
              additionalProperties: false
      responses:
        '200':
          description: Pool created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Uploaded successfully"
                  id:
                    type: string
                    example: "A1B2C3"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Bad Request"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Malformed API Key provided"
        '404':
          description: Topology not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Not Found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"

    get:
      summary: Get pools
      description: Get a specific pool or list all pools, with optional filtering for userIds or mainUsers
      tags:
        - Pool
      parameters:
        - name: poolId
          in: query
          schema:
            type: string
          description: ID of the pool to retrieve
          example: "ABC123"
        - name: userIds
          in: query
          schema:
            type: string
            enum: ["true"]
          description: If set to "true", returns only userIds for the specified pool (requires poolId)
          example: "true"
        - name: mainUsers
          in: query
          schema:
            type: string
            enum: ["true"]
          description: If set to "true", returns only mainUsers for the specified pool (requires poolId, cannot be used with userIds)
          example: "true"
      responses:
        '200':
          description: Pool data or list of pools
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: Single pool (full data)
                    properties:
                      poolId:
                        type: string
                        example: "ABC123"
                      createdBy:
                        type: string
                        example: "admin"
                      createdAt:
                        type: string
                        format: date-time
                        description: When the pool was created
                        example: "2025-08-30T14:30:45Z"
                      type:
                        type: string
                        enum: ["SHARED", "INDIVIDUAL"]
                        example: "SHARED"
                      topologyId:
                        type: string
                        example: "H4tCgb"
                      usersAndTeams:
                        type: array
                        items:
                          type: object
                          properties:
                            user:
                              type: string
                              example: "Alice Dan Mayers"
                            userId:
                              type: string
                              example: "BATCHalicedanmayers"
                            team:
                              type: string
                              example: "Red Team"
                            mainUserId:
                              type: string
                              example: "instructor1"
                              description: "Present only for SHARED pools"
                      note:
                        type: string
                        example: "Training session"
                      ctfdData:
                        type: boolean
                        description: Indicates if CTFD data is available for this pool
                        example: true
                  - type: object
                    description: Single pool (userIds only)
                    properties:
                      poolId:
                        type: string
                        example: "ABC123"
                      userIds:
                        type: array
                        items:
                          type: string
                        example: ["BATCHalicedanmayers", "BATCHbobsmith"]
                  - type: object
                    description: Single pool (mainUsers only)
                    properties:
                      poolId:
                        type: string
                        example: "ABC123"
                      mainUsers:
                        type: array
                        items:
                          type: string
                        example: ["instructor1", "instructor2"]
                  - type: array
                    description: List of all pools
                    items:
                      type: object
                      properties:
                        poolId:
                          type: string
                          example: "ABC123"
                        createdBy:
                          type: string
                          example: "admin"
                        createdAt:
                          type: string
                          format: date-time
                          description: When the pool was created
                          example: "2025-08-30T14:30:45Z"
                        type:
                          type: string
                          enum: ["SHARED", "INDIVIDUAL"]
                          example: "SHARED"
                        topologyId:
                          type: string
                          example: "H4tCgb"
                        note:
                          type: string
                          example: "Training session"
                        ctfdData:
                          type: boolean
                          description: Indicates if CTFD data is available for this pool
                          example: true
        '400':
          description: Bad Request - Invalid parameters (e.g., both userIds and mainUsers specified)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Bad Request"
        '404':
          description: Pool not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Not Found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"

    delete:
      summary: Delete a pool
      description: Delete a specific pool and all its data
      tags:
        - Pool
      parameters:
        - name: poolId
          in: query
          required: true
          schema:
            type: string
          description: ID of the pool to delete
          example: "ABC123"
      responses:
        '204':
          description: Pool deleted successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Pool not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pool/topology:
    patch:
      summary: Update pool topology
      description: Update the topology ID for a specific pool.
      tags:
        - Pool
      parameters:
        - in: query
          name: poolId
          schema:
            type: string
          required: true
          description: Pool ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                topologyId:
                  type: string
                  pattern: "^[a-zA-Z0-9]{6}$"
              required:
                - topologyId
      responses:
        '200':
          description: Pool topology updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Updated successfully"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Not Found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"
  /pool/note:
    patch:
      summary: Update pool note
      description: Update the note for a specific pool.
      tags:
        - Pool
      parameters:
        - in: query
          name: poolId
          schema:
            type: string
          required: true
          description: Pool ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  maxLength: 15
              required:
                - note
      responses:
        '200':
          description: Pool note updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Updated successfully"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Not Found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"
  /pool/users:
    post:
      summary: Check if user IDs exist in pools
      description: Check if specific user IDs already exist in any pool across the system
      tags:
        - Pool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  example: ["BATCHuser1", "BATCHuser2", "BATCHalicedanmayers"]
              required:
                - userIds
              additionalProperties: false
      responses:
        '200':
          description: User ID check results
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    userId:
                      type: string
                      example: "BATCHuser1"
                    exists:
                      type: boolean
                      description: Whether this user ID exists in any pool
                      example: true
        '400':
          description: Bad Request - Invalid request body or validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Add users to pool
      description: Add new users to an existing pool, combining them with existing users. Pool type determines mainUserId requirements.
      tags:
        - Pool
      parameters:
        - name: poolId
          in: query
          required: true
          schema:
            type: string
            pattern: "^[a-zA-Z0-9]{6}$"
          description: Pool ID to add users to
          example: "ABC123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - usersAndTeams
              properties:
                type:
                  type: string
                  enum: ["SHARED", "INDIVIDUAL"]
                  description: Pool type - must match existing pool type
                  example: "SHARED"
                usersAndTeams:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - user
                    properties:
                      user:
                        type: string
                        description: Full name of the user
                        example: "Alice Dan Mayers"
                      team:
                        type: string
                        description: Team name (optional, but if one user has a team, all must have teams)
                        example: "Red Team"
                      mainUserId:
                        type: string
                        description: Main user ID (required for SHARED pools, not allowed for INDIVIDUAL pools, must match existing mainUserIds when adding to SHARED pool)
                        example: "instructor1"
                  description: List of users to add to the pool
              additionalProperties: false
      responses:
        '200':
          description: Users added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Users added successfully"
        '400':
          description: Bad Request - Invalid request body, duplicate users, team consistency violation, mainUserId mismatch, or users already exist in other pools
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Bad Request"
        '404':
          description: Pool Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Not Found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"

  # Ctfd Data
  /ctfd/data:
    get:
      summary: Get flag data
      tags:
        - Ctfd Flag Data
      parameters:
        - in: query
          name: poolId
          schema:
            type: string
          required: true
          description: Pool ID to get CTFD data for
          example: "DEF456"
      responses:
        '200':
          description: CTFD data found
          content:
            application/json:
              schema:
                type: object
                properties:
                  ctfdData:
                    type: array
                    items:
                      type: object
                      properties:
                        user:
                          type: string
                          example: "alice"
                        password:
                          type: string
                          example: "password123"
                        team:
                          type: string
                          example: "Red Team"
                        flags:
                          type: array
                          items:
                            type: object
                            properties:
                              variable:
                                type: string
                                example: "FLAG_1"
                              contents:
                                type: string
                                example: "example_flag_content"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Pool not found or CTFD data not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Ctfd Flag Data
      summary: Extract and retrieve CTFd flags from user logs
      description: Extracts flags from deployed range logs for all users in a pool and saves the data to ctfd_data.json
      security:
        - ApiKeyAuth: []
      parameters:
        - name: poolId
          in: query
          required: true
          description: The pool ID to extract flags for
          schema:
            type: string
      responses:
        '200':
          description: Flags extracted and saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Flags extracted and saved successfully"
                  poolId:
                    type: string
                    example: "pool123"
                  ctfd_data:
                    type: array
                    items:
                      type: object
                      properties:
                        user:
                          type: string
                          example: "user1"
                        password:
                          type: string
                          example: "password123"
                        team:
                          type: string
                          example: "team1"
                        flags:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                              value:
                                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
        '401':
          description: Unauthorized - Invalid or missing API key
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
        '404':
          description: Not Found - Pool not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Not Found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"
  /ctfd/data/logins:
    get:
      summary: Get CTFd logins
      description: Retrieve user login information in CSV format for a specific CTFd data ID or all available data.
      tags:
        - Ctfd Flag Data
      parameters:
        - in: query
          name: poolId
          schema:
            type: string
            pattern: "^[a-zA-Z0-9]{6}$"
          required: false
          description: Data ID to retrieve logins for. If not provided, retrieves logins for all data.
      responses:
        '200':
          description: Logins retrieved successfully
          content:
            text/plain:
              schema:
                type: string
                example: |
                  user1, password1, team1
                  user2, password2, team2
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Not Found"

  # Ludus User Management
...